requirements.txt
streamlit
tensorflow
pillow
opencv-python-headless
numpy
huggingface-hub
